Fix initial segfaulting and test files
Changes made in TreeBuilderManager.cpp to fix probable memory alignment issues with the Distance Calculator. This fixed the initial segfaulting in NewCalcDNA (but there is more to figure out.)
Also made changes to the fixtures - random.faa changed to amino.faa to indicate it has amino acid sequences in it. Generated and added DNA.fa, for testing the DNA functions we've been working on.
Program is currently producing a tree with amino.faa but segfaulting at src/DistanceCalculatorAVX.cpp:368 with the dna file.