Fix all segfaults and produce a DNA tree
Made two major changes that fixed the segfaulting issue on DNA
sequences and allow it to produce an (albeit not quite correct still)
tree.
NewCalcDNA now explicitly loads a chunk from the character arrays into
seq1/seq2. Was able to change it from being an unaligned load by
changing the allocation of convertedSequences in ConvertAllDNA to use
aligned_alloc instead of new.