Avx debug (#25)
* extraction stage changes finished. Result not yet confirmed, but it is able to compile now!

* Still debugging, this is just a safety push

* Fix initial segfaulting and test files

Changes made in TreeBuilderManager.cpp to fix probable memory alignment issues with the Distance Calculator. This fixed the initial segfaulting in NewCalcDNA (but there is more to figure out.)
Also made changes to the fixtures - random.faa changed to amino.faa to indicate it has amino acid sequences in it. Generated and added DNA.fa, for testing the DNA functions we've been working on.
Program is currently producing a tree with amino.faa but segfaulting at src/DistanceCalculatorAVX.cpp:368 with the dna file.

* Fix all segfaults and produce a DNA tree

Made two major changes that fixed the segfaulting issue on DNA
sequences and allow it to produce an (albeit not quite correct still)
tree.
NewCalcDNA now explicitly loads a chunk from the character arrays into
seq1/seq2. Was able to change it from being an unaligned load by
changing the allocation of convertedSequences in ConvertAllDNA to use
aligned_alloc instead of new.

* Skip clang-format test

Clang-format seems to be having issues, we're skipping those tests for
now so the relevant lines in the makefile have been commented out.

* Complete the makefile

Somehow a large chunk of the makefile that exists in the AVX branch was
missing from this debug branch. This makefile has been updated to add
the missing parts, with check-format still being skipped.

* Fix makefile echo issue

Small formatting issue in check-format echo statement fixed so test
will actually be skipped now... for real.

* Make modifications for PR feedback

Old, commented out seq/gap initialization marked as previous versions
Comment on gather stage clarified to denote separate lanes
Deleting the empty mistake of a test file