Changes to attempt bitpacking debug, add "bug notes"
Tracked down bug in bitpacking, so most of the work done here reflects tracking that down (many print statements) and some changes to the convertAllDNA function were made to attempt to fix it. It's still a problem - comments marked BUG NOTE were added to document what's going on and try to demonstrate why it's a more extensive problem than originally anticipated.