Change file structure and add generation script & first fixtures
Changed the base file structure of the project (removing the redundant "NINJA" folder) and putting in src and a subfolder, fixtures, as we had it previously on the AVX branch.

Added the generation script I wrote, now adapted to generate a 3 sequence alignment and the subsequent distance matrix. This script, and the files we make from it, will be what we use for testing - to be compared to NINJA output for accuracy.