Add test files of lengths previously discussed
Generated and added the length 10nt, 64nt, and 97nt sequence alignments as previously discussed. Comments on the bottom of each fasta file show the distance matrix, to compare NINJA output to later, and then the final line shows the input number of transitions and transversions chosen as input to the ninja-gen script, in case that info is needed later.