Convert to CMake project, change folder structure, begin testing
Converted project in its entirety from a basic makefile to a cmake project. I believe all off the pieces from the original makefile have been addressed in the distributed cmakelists.txt files, and it built and runs properly with the testing I did (DNA distance matrix), but it is worth double checking. Directions for building and running have been added to makefile - definitely not how/where it should permanently be documented, but it suffices for now.
Folder structure has also been changed, partially to accommodate that and partially to accommodate implementation of testing. The lib folder holds the google test suites repo. Testing has just begun implementation - the testing.cpp file will eventually be deleted, but it needs to be fully converted and distributed in the distance-matrix test file first.
Distance-matrix implementation is not currently working, and nowhere near complete, but it is a good start. Still figuring out how to access just the actual matrix returned from DistanceCalculator to compare to my generated output.