DistanceCalculator Test Implemented
I believe the unit test is implemented and *should* work (fingers crossed - it definitely compiles though!), but I haven't actually been able to confirm that. Right now we cannot run NINJA_test because the makefile needs to be modified to add the rule for a test target in addition to the all, debug, and clean rules. The actual calculation part has been separated from the sequence generation function, although I realized it was easier for the sequence generation function to call the kimura calculation function instead of separating them completely.
The testing.cpp file (which was the original script written to generate random sequences and give the correct distances between sequences) is still included just in case the reference is needed, but has been moved from the test folder up to the base directory - can be deleted once George confirms he doesn't need it!
Otherwise... all that's left from here (as far as directly comparing the distance matrix the DistanceCalculator class calculates to known correct output) is updating the makefile to allow the test build to happen, and confirm that the unit test can pass successfully.